scripts/requirements.txt
supabase
google-api-python-client
google-auth
google-auth-oauthlib
python-dotenv
